requirements.txt
torch==2.9.0
torchvision==0.24.0
streamlit==1.50.0
Pillow==10.3.0
scikit-learn==1.6.0
matplotlib==3.9.1
tqdm==4.66.5
pandas==2.2.3
numpy==1.26.4